requirements.txt
kafka-python
elasticsearch
grpcio
